requirements.txt
attrdict==2.0.1
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bitarray==0.9.0
bluetool==0.2.3
cached-property==1.5.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coloredlogs==10.0
dbus-python==1.2.8
decorator==4.4.0
enum34==1.1.6
Flask==1.0.2
gattlib==0.20150805
humanfriendly==4.18
idna==2.8
ipdb==0.12
ipython==5.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
lru-cache==0.2.3
MarkupSafe==1.1.1
monotonic==1.5
ordereddict==1.1
pathlib2==2.3.3
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==1.0.16
ptyprocess==0.6.0
PyBluez==0.22
pycairo==1.18.1
Pygments==2.3.1
PyGObject==3.32.1
PyOBEX==0.27
python-dateutil==2.8.0
requests==2.21.0
scandir==1.10.0
simplegeneric==0.8.1
six==1.12.0
tcpbridge==1.1.1
traitlets==4.3.2
urllib3==1.24.2
vobject==0.9.6.1
wcwidth==0.1.7
Werkzeug==0.15.2
